move preprocess to file
