+++ training_pipeline_edna_gpu.py
@@ -21,6 +21,8 @@ from subprocess import Popen
 import threading
 import logging
 
+import infer.modules.train.training_pipeline as pipeline #import preprocess_dataset
+
 # config = Config()
 
 # Taken from Config defaults
@@ -144,11 +146,7 @@ logger = logging.getLogger(__name__)
 
 #############################################################################
 
-sr_dict = {
-    "32k": 32000,
-    "40k": 40000,
-    "48k": 48000,
-}
+
 
 version_config_list = [
     "v1/32k.json",
@@ -176,82 +174,6 @@ def if_done(done, p):
             break
     done[0] = True
 
-#from infer-web.py
-
-# def make_dirs(exp_dir):
-#     #self.exp_dir = exp_dir
-
-#     gt_wavs_dir = "%s/0_gt_wavs" % exp_dir
-#     wavs16k_dir = "%s/1_16k_wavs" % exp_dir
-#     os.makedirs(exp_dir, exist_ok=True)
-#     os.makedirs(gt_wavs_dir, exist_ok=True)
-#     os.makedirs(wavs16k_dir, exist_ok=True)
-
-def preprocess_dataset(trainset_dir, exp_dir, sr, n_p, 
-                       config_vars = None, 
-                       now_dir = None,
-                       logger = None):
-    sr = sr_dict[sr]
-    print('------------------  Enter preprocess')
-    
-    print(f'now_dir: {now_dir}')
-
-    print('Making dirs')
-
-    #real_exp_dir = "%s/logs/%s" % (now_dir, exp_dir)
-
-    real_exp_dir = f"{now_dir}/logs/{exp_dir}"
-    gt_wavs_dir =  f"{now_dir}/logs/{exp_dir}/0_gt_wavs" 
-    wavs16k_dir =  f"{now_dir}/logs/{exp_dir}/1_16k_wavs"
-
-    print(f'Creating real_exp_dir dir: {real_exp_dir}')
-    print(f'Creating gt_wavs_dir dir: {gt_wavs_dir}')
-    print(f'Creating wavs16k_dir dir: {wavs16k_dir}')
-
-    #os.makedirs("%s/logs/%s" % (now_dir, exp_dir), exist_ok=True)   
-    os.makedirs(real_exp_dir, exist_ok=True) 
-    os.makedirs(gt_wavs_dir, exist_ok=True)
-    os.makedirs(wavs16k_dir, exist_ok=True)
-
-    f = open("%s/logs/%s/preprocess.log" % (now_dir, exp_dir), "w")
-    f.close()
-
-    cmd = '"%s" infer/modules/train/preprocess.py "%s" %s %s "%s/logs/%s" %s %.1f' % (
-        config_vars['python_cmd'],
-        trainset_dir,
-        sr,
-        n_p,
-        now_dir,
-        exp_dir,
-        config_vars['noparallel'],
-        config_vars['preprocess_per'],
-    )
-    
-    logger.info("Execute: " + cmd)
-    p = Popen(cmd, shell=True)
-    p.communicate()
-
-    # threading stuff 
-    # 煞笔gr, popen read都非得全跑完了再一次性读取, 不用gr就正常读一句输出一句;只能额外弄出一个文本流定时读
-    
-    # done = [False]
-    # threading.Thread(
-    #     target=if_done,
-    #     args=(
-    #         done,
-    #         p,
-    #     ),
-    # ).start()
-    # while 1:
-    #     with open("%s/logs/%s/preprocess.log" % (now_dir, exp_dir), "r") as f:
-    #         yield (f.read())
-    #     sleep(1)
-    #     if done[0]:
-    #         break
-    # with open("%s/logs/%s/preprocess.log" % (now_dir, exp_dir), "r") as f:
-    #     log = f.read()
-    # logger.info(log)
-    # yield log
 
 def extract_f0_feature(gpus, n_p, f0method, if_f0, exp_dir, version, gpus_rmvpe):
     gpus = gpus.split("-")
@@ -573,7 +495,7 @@ def train_index(exp_dir1, version19):
     # infos.append("成功构建索引，added_IVF%s_Flat_FastScan_%s.index"%(n_ivf,version19))
     #yield "\n".join(infos)
 
-preprocess_dataset(param_dict['trainset_dir'], 
+pipeline.preprocess_dataset(param_dict['trainset_dir'], 
                     param_dict['exp_dir'], 
                     param_dict['sr'], 
                     param_dict['num_proc'],
